requirements.txt
uvicorn==0.26.0
fastapi==0.109.0
python-dotenv==1.0.0